feat: value separate

+++ src/cache/cache.rs
@@ -2,4 +2,4 @@ use crate::block::builder::Block;
 use std::sync::Arc;
 
 // (sst id, block id)
-pub type BlockCache = moka::sync::Cache<(usize, usize), Arc<Block>>;
+pub type BlockCache = moka::sync::Cache<(u32, usize), Arc<Block>>;
